+++ spec/features/view_supplier_details_spec.rb
@@ -4,14 +4,15 @@ RSpec.feature 'Viewing a supplier' do
   let!(:supplier) { FactoryBot.create(:supplier, name: 'Test Supplier Ltd') }
   let!(:framework) { FactoryBot.create(:framework, name: 'Test Framework', short_name: 'RM0000') }
   let!(:agreement) { FactoryBot.create(:agreement, supplier: supplier, framework: framework) }
+  let!(:user) { FactoryBot.create(:user, name: 'User One', email: 'email_one@ccs.co.uk', suppliers: [supplier]) }
 
   before { sign_in_as_admin }
 
-  scenario 'shows paginated list of frameworks the supplier has an agreement for' do
+  scenario 'shows paginated list of agreements the supplier has an agreement for' do
     visit admin_supplier_path(supplier)
     expect(page).to have_content 'Test Supplier Ltd'
     expect(page).to have_content 'RM0000 Test Framework'
-    expect(page).to have_content 'Displaying 1 framework'
+    expect(page).to have_content 'Displaying 1 agreement'
   end
 
   scenario 'shows paginated list of the supplier’s tasks' do
@@ -28,7 +29,7 @@ RSpec.feature 'Viewing a supplier' do
       :task, period_month: 12, period_year: 2018, supplier: supplier, framework: framework, status: :in_progress
     )
     FactoryBot.create(
-      :submission_with_validated_entries, supplier: supplier, framework: framework, task: task
+      :submission_with_validated_entries, supplier: supplier, framework: framework, task: task, created_by: user
     )
 
     visit admin_supplier_path(supplier)
@@ -40,7 +41,8 @@ RSpec.feature 'Viewing a supplier' do
       :task, period_month: 12, period_year: 2018, supplier: supplier, framework: framework, status: :completed
     )
 
-    FactoryBot.create(:submission, supplier: supplier, framework: framework, task: task) do |submission|
+    FactoryBot.create(:submission, supplier: supplier, framework: framework, task: task,
+created_by: user) do |submission|
       FactoryBot.create(:submission_file, :with_attachment, submission: submission, filename: 'test.xlsx')
     end
 
@@ -53,8 +55,6 @@ RSpec.feature 'Viewing a supplier' do
   end
 
   scenario 'shows paginated list of the users linked to the supplier' do
-    FactoryBot.create(:user, suppliers: [supplier])
-
     visit admin_supplier_path(supplier)
     expect(page).to have_content 'Active?'
     expect(page).to have_content 'Displaying 1 user'
@@ -68,12 +68,11 @@ framework: framework)
       FactoryBot.create(:user, suppliers: [supplier])
     end
     FactoryBot.create(:task, period_month: 12, period_year: 2017, supplier: supplier, framework: framework)
-    FactoryBot.create(:user, suppliers: [supplier])
 
     visit admin_supplier_path(supplier)
     click_link('Next »', match: :first)
     expect(page).to have_content 'Displaying task 13 - 13 of 13 in total'
     expect(page).to have_content 'Displaying users 1 - 12 of 13 in total'
-    expect(page).to have_content 'Displaying 1 framework'
+    expect(page).to have_content 'Displaying 1 agreement'
   end
 end